agents/agent1_student/requirements.txt
fastapi
uvicorn
langchain
langchain-community
qdrant-client
python-docx
PyPDF2
